PyTorch/computer_vision/classification/lightning/resnet/requirements.txt
lightning==2.0.4
lightning-habana==1.0.0
torchmetrics==0.10.0
